Move more logic to library
