+++ libs/skills/package.json
@@ -20,6 +20,7 @@
   "peerDependencies": {
     "@epam/ai-dial-catalog": "*",
     "@epam/ai-dial-chat-shared": "*",
+    "@epam/ai-dial-conversation-input": "*",
     "@epam/ai-dial-ui-kit": "*",
     "@tabler/icons-react": "^3.0.0",
     "react": "^19.0.0"
