+++ libs/skills/src/index.ts
@@ -1,5 +1,7 @@
 export { FavoriteSkillsPanel } from './components/FavoriteSkillsPanel/FavoriteSkillsPanel';
+export { SkillCatalogModal } from './components/SkillCatalogModal/SkillCatalogModal';
 export { SkillDetailsSidePanel } from './components/SkillDetailsSidePanel/SkillDetailsSidePanel';
+export { useSkillSelectorOverlay } from './hooks/useSkillSelectorOverlay/useSkillSelectorOverlay';
 export { buildFavoriteSkillItem } from './models/favorite-skill-item';
 export type {
   FavoriteSkillItem,
@@ -10,4 +12,11 @@ export type {
   FavoriteSkillsPanelLabels,
   FavoriteSkillsPanelProps,
 } from './models/favorite-skills-panel-props';
+export type { SkillCatalogModalProps } from './models/skill-catalog-modal-props';
 export type { SkillDetailsSidePanelProps } from './models/skill-details-side-panel-props';
+export type {
+  SkillDetailsPanelComponentProps,
+  SkillSelectorOverlayLabels,
+  UseSkillSelectorOverlayOptions,
+  UseSkillSelectorOverlayResult,
+} from './models/skill-selector-overlay';
