+++ libs/skills/tsconfig.lib.json
@@ -24,6 +24,9 @@
   ],
   "references": [
     {
+      "path": "../conversation-input/tsconfig.lib.json"
+    },
+    {
       "path": "../chat-shared/tsconfig.lib.json"
     },
     {
